+++ src/components/Notification/Notification.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import Dismiss from 'components/Dismiss';
 import './Notification.scss';
 
 const Notification = ({ type, title, message, onDismissClick }) => {
-	const msgColorClass = type ? `is-${type}` : '';
 	return (
 		<React.Fragment>
 			<li className="notification-item">
 				<div className={`notification ${type || ''}`}>
-					<button className="delete" aria-label="dismiss" onClick={onDismissClick}></button>
+					<Dismiss onClick={onDismissClick} />
 					<h2 className="title is-4">{title}</h2>
 					<p>{message}</p>
 				</div>